backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12

# Excel
polars==1.9.0
openpyxl==3.1.5
xlsxwriter==3.2.0

# Validation & settings
pydantic[email]==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# Utilities
python-dotenv==1.0.1
aiofiles==24.1.0

# Rate limiting
slowapi==0.1.9
